requirements.txt
bottle==0.12.13
cycler==0.10.0
flake8==3.3.0
matplotlib==2.0.2
mccabe==0.6.1
numpy==1.12.1
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
requests==2.12.4
